Create separate coding module for encodings
