+++ echopype/convert/set_groups_ek60.py
@@ -6,7 +6,8 @@ import numpy as np
 import xarray as xr
 from _echopype_version import version as ECHOPYPE_VERSION
 
-from .set_groups_base import DEFAULT_CHUNK_SIZE, SetGroupsBase, set_encodings
+from ..utils.coding import set_encodings
+from .set_groups_base import DEFAULT_CHUNK_SIZE, SetGroupsBase
 
 
 class SetGroupsEK60(SetGroupsBase):
